Change replacement_delay type from Decimal to float in BootstrapPMMConfig to prevent asyncio.sleep issues

+++ scripts/bootstrap_pmm.py
@@ -33,7 +33,7 @@ class BootstrapPMMConfig(BaseClientModel):
     randomize_order_amount: bool = Field(False)
     replace_all_every_increment: bool = Field(False)  # Turn on replace all orders ever n seconds
     replacement_increment : float = Field(10.0)  # Amount in seconds to have all orders be replaced.
-    replacement_delay : Decimal = Field(1.5)  # Amount of time to wait between replacing each order
+    replacement_delay : float = Field(1.5)  # Amount of time to wait between replacing each order
     random_order_floor: Decimal = Field(1)  # in quote currency (USDT)
     random_order_ceiling: Decimal = Field(2)  # in quote currency (USDT)
     price_ceiling: Decimal = Field(2.0)
